feat: jwt 인증 추가

+++ controller/LikeController.js
@@ -1,38 +1,77 @@
+const jwt = require("jsonwebtoken");
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
+const dotenv = require("dotenv");
+
+dotenv.config();
 
 const addLike = (req, res) => {
-  const { id } = req.params; // book_id
-  const { user_id } = req.body;
+  const book_id = req.params.id;
 
-  let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
-  let values = [user_id, id];
+  let authorization = ensureAuthorization(req);
 
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
+    let values = [authorization.id, book_id];
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
-  const { id } = req.params; // book_id
-  const { user_id } = req.body;
+  const book_id = req.params.id;
 
-  let sql = "DELETE FROM likes WHERE user_id=? AND liked_book_id=?;";
-  let values = [user_id, id];
+  let authorization = ensureAuthorization(req);
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다.",
+    });
+  } else {
+    let sql = "DELETE FROM likes WHERE user_id=? AND liked_book_id=?;";
+    let values = [authorization.id, book_id];
 
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
+
+// jwt 토큰 꺼내서 복호화 시켜주는 함수
+function ensureAuthorization(req) {
+  try {
+    let receivedJwt = req.headers["authorization"];
+    let decodedJwt = jwt.verify(receivedJwt, process.env.SECRET_KEY);
+    return decodedJwt;
+  } catch (err) {
+    console.log(err.name);
+    console.log(err.message);
+
+    return err;
+  }
+}
 
 module.exports = { addLike, removeLike };
